Fix syntax and type errors in mobile build

+++ mobile_app/lib/screens/product_details_screen.dart
@@ -33,7 +33,7 @@ class _ProductDetailsScreenState extends ConsumerState<ProductDetailsScreen> {
             quantity = product.moqKg;
             // if stock is lower than moq, we set to stock, or 0 if out of stock
             if (quantity > product.availableStockKg) {
-              quantity = product.availableStockKg;
+              quantity = product.availableStockKg.toInt();
             }
             isInitialized = true;
           }
